Rename `actions/notifications` to `actions/emergencyNotifications`
